+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - (**BREAKING**) Removed `TypedBuilder` implementation for `AqlQuery`, replaced it with an explicit builder pattern
+- (**BREAKING**) `Collection::load` and `Collection::unload` required `arango3_7` feature
 - Clippy improvements
 - Added `inline` hints
 - Added `must_use` attributes